feat: function to get loan callouts added to loanutils file
